+++ src/Settings/NumberGeneratorSettings/index.js
@@ -1,0 +1,1 @@
+export { NumberGeneratorSettingsForm } from './NumberGeneratorSettingsForm';